♻️(backend) regroup common live session test method
Regroup the commonly used test methods for the LiveSession
API test into a `base` module.
This was missed during the split with gitfilesplit so the
history is lost for this specific method, I think it is not
an issue.
